feat: split old and new API so that we can access them with ngx-sub-form and ngx-sub-form/new to avoid making a breaking change
