+++ projects/ngx-sub-form/new/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/ngx-sub-form",
+  "lib": {
+    "entryFile": "../public_api.new.ts"
+  },
+  "whitelistedNonPeerDependencies": ["fast-deep-equal"]
+}
