+++ projects/ngx-sub-form/new/src/ngx-sub-form.ts
@@ -1,9 +1,9 @@
-import { ɵivyEnabled, ɵmarkDirty as markDirty } from '@angular/core';
+import { ɵmarkDirty as markDirty } from '@angular/core';
 import isEqual from 'fast-deep-equal';
 import { decorateObservableLifecycle, getObservableLifecycle } from 'ngx-observable-lifecycle';
 import { EMPTY, forkJoin, Observable, of } from 'rxjs';
 import { delay, filter, map, mapTo, shareReplay, switchMap, take, takeUntil, tap } from 'rxjs/operators';
-import { isNullOrUndefined } from '../ngx-sub-form-utils';
+import { isNullOrUndefined } from '../../src/lib/shared/ngx-sub-form-utils';
 import {
   createFormDataFromOptions,
   getControlValueAccessorBindings,
