+++ projects/ngx-sub-form/src/lib/deprecated/ngx-root-form.component.spec.ts
@@ -3,8 +3,8 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormArray, FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 import { BehaviorSubject } from 'rxjs';
+import { ArrayPropertyKey, ArrayPropertyValue, Controls } from '../shared/ngx-sub-form-utils';
 import { NgxRootFormComponent } from './ngx-root-form.component';
-import { ArrayPropertyKey, ArrayPropertyValue, Controls } from './ngx-sub-form-utils';
 import { DataInput } from './ngx-sub-form.decorators';
 import { NgxFormWithArrayControls } from './ngx-sub-form.types';
