+++ projects/ngx-sub-form/src/public_api.ts
@@ -2,9 +2,9 @@
  * Public API Surface of sub-form
  */
 
-export * from './lib/ngx-sub-form-utils';
-export * from './lib/ngx-sub-form.component';
-export * from './lib/ngx-root-form.component';
-export * from './lib/ngx-automatic-root-form.component';
-export * from './lib/ngx-sub-form.types';
-export * from './lib/ngx-sub-form.decorators';
+export * from './lib/deprecated/ngx-automatic-root-form.component';
+export * from './lib/deprecated/ngx-root-form.component';
+export * from './lib/deprecated/ngx-sub-form.component';
+export * from './lib/deprecated/ngx-sub-form.decorators';
+export * from './lib/shared/ngx-sub-form-utils';
+export * from './lib/shared/ngx-sub-form.types';
