+++ src/app/main-rewrite/listing/listing-form/droid-listing/astromech-droid/astromech-droid.component.ts
@@ -1,8 +1,7 @@
 import { Component, forwardRef } from '@angular/core';
 import { FormControl, Validators } from '@angular/forms';
 import { subformComponentProviders } from 'ngx-sub-form';
-import { createForm, NgxSubForm } from '../../../../../../../projects/ngx-sub-form/src/lib/new/ngx-sub-form';
-import { FormType } from '../../../../../../../projects/ngx-sub-form/src/lib/new/ngx-sub-form.types';
+import { createForm, FormType, NgxSubForm } from 'ngx-sub-form/new';
 import { AstromechDroid, AstromechDroidShape, DroidType } from '../../../../../interfaces/droid.interface';
 
 @NgxSubForm()
