+++ src/app/main-rewrite/listing/listing-form/droid-listing/protocol-droid/protocol-droid.component.ts
@@ -2,8 +2,7 @@ import { Component, forwardRef } from '@angular/core';
 import { FormControl, Validators } from '@angular/forms';
 import { getObservableLifecycle } from 'ngx-observable-lifecycle';
 import { subformComponentProviders } from 'ngx-sub-form';
-import { createForm, NgxSubForm } from '../../../../../../../projects/ngx-sub-form/src/lib/new/ngx-sub-form';
-import { FormType } from '../../../../../../../projects/ngx-sub-form/src/lib/new/ngx-sub-form.types';
+import { createForm, FormType, NgxSubForm } from 'ngx-sub-form/new';
 import { DroidType, Languages, ProtocolDroid } from '../../../../../interfaces/droid.interface';
 
 @NgxSubForm()
