+++ tsconfig.json
@@ -21,7 +21,7 @@
     "lib": ["es2018", "dom"],
     "paths": {
       "ngx-sub-form": ["projects/ngx-sub-form/src/public_api"],
-      "ngx-sub-form/*": ["projects/ngx-sub-form/src/public_api"]
+      "ngx-sub-form/new": ["projects/ngx-sub-form/public_api.new"]
     }
   }
 }
